// src/components/types/package-item.ts
export type PackageItem = {
  uuid_package: string;
  condominium_name: string;
  blockReceiver?: string;
  apartmentReceiver?: string;
  blockOwner?: string;
  apartmentOwner?: string;
  receiverName?: string;
  ownerName?: string;
  created_at: string;
  note?: string;
  status_package: string;
  confirmation_code: string;
};